api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
reportlab
openpyxl
python-docx
